fix en float de costos

+++ src/helpers/procedures.ts
@@ -275,7 +275,7 @@ export const updateProcedureCost = async (id: string, newCost: string): Promise<
   try {
     client.query('BEGIN');
     const res = (await client.query(queries.GET_UTMM_VALUE)).rows[0]
-    const response = (await client.query(queries.UPDATE_PROCEDURE_COST, [id, newCost, parseInt(newCost) * parseInt(res.valor_en_bs)])).rows[0];
+    const response = (await client.query(queries.UPDATE_PROCEDURE_COST, [id, newCost, parseFloat(newCost) * parseFloat(res.valor_en_bs)])).rows[0];
     const newProcedure = (await client.query(queries.GET_ONE_PROCEDURE, [id])).rows[0];
     const procedure: Partial<TipoTramite> = {
       id: newProcedure.id_tipo_tramite,
